pyproject.toml
[tool.poetry]
name = "image_dist"
version = "0.1.0"
description = ""
authors = ["dlee <derekclee323@g.ucla.edu>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.0.1"
PyDrive = "^1.3.1"
OSMPythonTools = "^0.2.9"
numpy = "^1.19.4"
requests = "^2.25.1"
haversine = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
